[MS-525] rename roc-wrapper to roc-v1
