+++ face/capture/build.gradle.kts
@@ -30,7 +30,7 @@ dependencies {
 
     //BIO SDK
     implementation(project(":face:infra:base-bio-sdk"))
-    implementation(project(":face:infra:roc-wrapper"))
+    implementation(project(":face:infra:roc-v1"))
 
     implementation(libs.androidX.cameraX.core)
     implementation(libs.androidX.cameraX.lifecycle)
