+++ face/infra/roc-v1/README.md
@@ -1,4 +1,4 @@
-# ROC Wrapper
+# ROC V1
 
 This module provides a wrapper for accessing all ROC SDK APIs.
 It implements all operations in `infra:face-bio-sdk module` and provides the implemented classes in `RocWrapperModule`.
