+++ face/infra/roc-v1/src/main/AndroidManifest.xml
