+++ face/infra/roc-v1/src/main/java/com/simprints/face/infra/rocv1/RocV1WrapperModule.kt
@@ -1,11 +1,11 @@
-package com.simprints.infra.rocwrapper
+package com.simprints.face.infra.rocv1
 
 import com.simprints.infra.facebiosdk.detection.FaceDetector
 import com.simprints.infra.facebiosdk.initialization.FaceBioSdkInitializer
 import com.simprints.infra.facebiosdk.matching.FaceMatcher
-import com.simprints.infra.rocwrapper.detection.RankOneFaceDetector
-import com.simprints.infra.rocwrapper.initialization.RankOneInitializer
-import com.simprints.infra.rocwrapper.matching.RankOneFaceMatcher
+import com.simprints.face.infra.rocv1.detection.RankOneFaceDetector
+import com.simprints.face.infra.rocv1.initialization.RankOneInitializer
+import com.simprints.face.infra.rocv1.matching.RankOneFaceMatcher
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -13,7 +13,7 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class RocWrapperModule {
+abstract class RocV1WrapperModule {
 
     @Binds
     abstract fun provideSdkInitializer(impl: RankOneInitializer): FaceBioSdkInitializer
