+++ face/infra/roc-v1/src/main/java/com/simprints/face/infra/rocv1/initialization/RankOneInitializer.kt
@@ -1,4 +1,4 @@
-package com.simprints.infra.rocwrapper.initialization
+package com.simprints.face.infra.rocv1.initialization
 
 import android.app.Activity
 import com.simprints.core.ExcludedFromGeneratedTestCoverageReports
