+++ face/infra/roc-v1/src/test/java/com/simprints/infra/rocwrapper/matching/RankOneFaceMatcherTest.kt
@@ -1,6 +1,7 @@
 package com.simprints.infra.rocwrapper.matching
 
 import com.google.common.truth.Truth
+import com.simprints.face.infra.rocv1.matching.RankOneFaceMatcher
 import org.junit.Test
 
 class RankOneFaceMatcherTest {
